Move ApiClientError related tests to its module
